Basic Event Joining and Leaving work!
You can now Join and leave Events by reacting with certain emojis. When you react with one of the corresponding emojis, it will add you to the corresponding list. Then it updates the embed in the message using message.edit. In order to update the embed, a new RichEmbed must be created. The GenerateEventEmbed will take in an event and generate a new richEmbed that can be used to send. 

members are added to events using the messageReactionAdd and messageReactionRemove events. Whenever a reaction is added, Quartermaster checks to see if the reaction was applied to an event. If it was the reaction emoji name is put into a switch statement and based on the emoji it will run the code for adding or removing from a certain list. When adding  or removing from a list. Quarter master looks at all users who reacted with that emoji adds them to an array and updates the event list with the new array. Then updates the embeds. 

Known Issues: 
-You can be on multiple lists at the same time
-If you change you name while in a list. The event will not fix your name in event in unless you rejoin the list.
-Sometimes upon event creation when confirming the event, Quartermaster will imediately think you put in an invalid answer and re-ask for confirmation.